Revert "Debuging products (#108)"
This reverts commit 246a6919520748207e233c093793f8950dcd2abd.
