+++ client/.gitignore
@@ -21,6 +21,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-config/keys_dev.js
-.env
