--- client/src/components/Carousel/styles.css
@@ -1,36 +1,0 @@
-.swiper {
-  padding-bottom: 30px;
-}
-
-.swiper-slide {
-  display: flex;
-  justify-content: center;
-}
-
-.swiper-pagination {
-  display: flex;
-  flex-wrap: nowrap;
-  justify-content: center;
-}
-.swiper-pagination-fraction {
-  bottom: 0;
-}
-.swiper-pagination-bullet {
-  width: 25px;
-  height: 5px;
-  border-radius: 5px;
-}
-.swiper-pagination-bullet-active {
-  background: rgb(46, 45, 45);
-}
-
-.swiper-button-prev,
-.swiper-button-next {
-  color: rgb(202, 199, 199);
-}
-.swiper-button-prev:hover,
-.swiper-button-next:hover {
-  color: rgb(46, 45, 45);
-}
-
-/*# sourceMappingURL=styles.css.map */
