+++ client/src/components/ProductCard/index.jsx
@@ -2,24 +2,143 @@ import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
-import { Box, useMediaQuery } from '@mui/material';
-import { useTheme } from '@mui/system';
+import {
+  Card,
+  CardContent,
+  Typography,
+  Box,
+  useMediaQuery,
+} from '@mui/material';
+import { styled, useTheme } from '@mui/system';
 import SvgIcon from '@mui/material/SvgIcon';
 import { AdvancedImage } from '@cloudinary/react';
 import { minimumPad } from '@cloudinary/url-gen/actions/resize';
+import { byRadius } from '@cloudinary/url-gen/actions/roundCorners';
 import getImg from '../../cloudinary';
 import CartIcon from '../Icons/cartIcon/cartIcon';
+import { RadiusButton } from '../Buttons';
 import { setProduct } from '../../redux/slices/productSlice';
 import { addToBasket } from '../../redux/slices/basketSlice';
-import {
-  CardContainer,
-  DetailButton,
-  AddToCartBtn,
-  Label,
-  CardInfo,
-  ProductName,
-  ProductPrice,
-} from '../../themes/themeProductCard';
+
+const CardContainer = styled(Card)(() => ({
+  outline: `solid 1px transparent`,
+  borderRadius: '15px',
+  marginBottom: '5%',
+  height: '100%',
+}));
+
+const DetailButton = styled(RadiusButton)(({ theme }) => ({
+  fontSize: '0.625rem',
+  lineHeight: '0.75rem',
+  maxWidth: '104px',
+  maxHeight: '26px',
+  [theme.breakpoints.between('xs', 'md')]: {
+    textTransform: 'capitalize',
+    minWidth: '77px',
+    minHeight: '20px',
+  },
+  [theme.breakpoints.between('md', 'lg')]: {
+    fontSize: '0.75rem',
+    minWidth: '111px',
+    minHeight: '28px',
+  },
+  [theme.breakpoints.up('lg')]: {
+    fontSize: '0.625rem',
+    minWidth: '104px',
+    minHeight: '26px',
+  },
+}));
+
+const AddToCartBtn = styled(RadiusButton)(({ theme }) => ({
+  backgroundColor: 'white',
+  outline: `1px solid ${theme.palette.primary.buttonhover}`,
+  maxWidth: '40px',
+  maxHeight: '26px',
+  position: 'relative',
+  '&:hover': {
+    backgroundColor: theme.palette.primary.buttonhover,
+    outline: `1px solid transparent`,
+  },
+  '& svg': {
+    marginTop: '5%',
+    minWidth: '13px',
+    minHeight: '12px',
+    maxWidth: '18px',
+    maxHeight: '20px',
+  },
+  [theme.breakpoints.between('xs', 'md')]: {
+    minWidth: '29px',
+    minHeight: '20px',
+    '& svg': {
+      width: '17px',
+      height: '16px',
+    },
+  },
+  [theme.breakpoints.between('md', 'lg')]: {
+    minWidth: '42px',
+    minHeight: '29px',
+    '& svg': {
+      width: '18px',
+      height: '20px',
+    },
+  },
+  [theme.breakpoints.up('lg')]: {
+    minWidth: '40px',
+    minHeight: '26px',
+    '& svg': {
+      width: '17px',
+      height: '18px',
+    },
+  },
+}));
+
+const Label = styled(Typography)(({ theme }) => ({
+  color: theme.palette.primary.dark,
+  fontSize: '0.625rem',
+  lineHeight: '0.75rem',
+}));
+
+const CardInfo = styled(CardContent)(({ theme }) => ({
+  color: theme.palette.primary.light,
+  //   fontFamily: theme.typography.const.fontFamily.secondary,
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'baseline',
+  padding: '0',
+  paddingLeft: '8%',
+  marginTop: '3%',
+  width: '93%',
+  [theme.breakpoints.between('xs', 'sm')]: {
+    display: 'block',
+    textAlign: 'left',
+    paddingLeft: '5%',
+  },
+  '&:last-child': {
+    paddingBottom: '0',
+  },
+}));
+
+const ProductName = styled(Typography)(({ theme }) => ({
+  fontWeight: theme.typography.const.fontWeight.bold,
+  width: '70%',
+  textOverflow: 'ellipsis',
+  overflow: 'hidden',
+  [theme.breakpoints.between('xs', 'sm')]: {
+    fontSize: '.9rem',
+    lineHeight: '1rem',
+    marginBottom: '2%',
+  },
+  [theme.breakpoints.up('xl')]: {
+    fontSize: '.875rem',
+    lineHeight: '1rem',
+  },
+}));
+
+const ProductPrice = styled(Typography)(() => ({
+  //   fontWeight: '400',
+  fontSize: '.875rem',
+  lineHeight: '1rem',
+}));
 
 function ProductCard({ product }) {
   const dispatch = useDispatch();
@@ -96,7 +215,8 @@ function ProductCard({ product }) {
               minimumPad()
                 .width(getImageSize().width)
                 .height(getImageSize().height)
-            )}
+            )
+            .roundCorners(byRadius(15, 15))}
           alt={product.name + product.color}
         />
         <Box
@@ -168,7 +288,7 @@ ProductCard.propTypes = {
     imageUrls: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
     quantity: PropTypes.number.isRequired,
     color: PropTypes.string.isRequired,
-    brand: PropTypes.string,
+    brand: PropTypes.string.isRequired,
     storage: PropTypes.string,
     itemNo: PropTypes.string,
     description: PropTypes.string,
