+++ client/src/components/ProductsList/index.jsx
@@ -2,38 +2,51 @@ import { React, useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useLocation } from 'react-router-dom';
 import { Grid, Box, Pagination, LinearProgress } from '@mui/material';
-import { useTheme } from '@mui/system';
+import { styled, useTheme } from '@mui/system';
 import PropTypes from 'prop-types';
 import { fetchProducts } from '../../redux/slices/productsSlice';
 import ProductCard from '../ProductCard';
 import {
   categoriesFilter,
-  homePageProducts,
-  isFetchingAllProducts,
   isFetchingProductsList,
   maximalPrice,
   minimalPrice,
   productsList,
   totalNumberProducts,
 } from '../../redux/selectors';
-import { getAllHomeProducts } from '../../redux/slices/allProdsHomeSlice';
-import StyledGrid from '../../themes/themeProductsList';
+
+const StyledGrid = styled(Grid)(({ theme }) => ({
+  display: 'flex',
+  justifyContent: 'center',
+  height: 'auto',
+  [theme.breakpoints.between('xs', 'md')]: {
+    '&:nth-of-type(n+5)': {
+      display: 'none',
+    },
+  },
+  [theme.breakpoints.between('md', 'lg')]: {
+    '&:nth-of-type(n+7)': {
+      display: 'none',
+    },
+  },
+  [theme.breakpoints.up('lg')]: {
+    '&:nth-of-type(n+9)': {
+      display: 'none',
+    },
+  },
+}));
 
 function ProductsList({ perPage }) {
   const theme = useTheme();
   const [currentPage, setCurrentPage] = useState(1);
   const total = useSelector(totalNumberProducts);
-  const productsForProducts = useSelector(productsList);
-  const productsForHomePage = useSelector(homePageProducts);
+  const products = useSelector(productsList);
   const categories = useSelector(categoriesFilter);
   const minFilterPrice = useSelector(minimalPrice);
   const maxFilterPrice = useSelector(maximalPrice);
   const formattedMinPrice = minFilterPrice !== null ? minFilterPrice : 7;
   const formattedMaxPrice = maxFilterPrice !== null ? maxFilterPrice : 100000;
-  const isFetchingProducts = useSelector(isFetchingProductsList);
-  const isFetchingHomeProds = useSelector(isFetchingAllProducts);
-  const location = useLocation();
-  const currentPath = location.pathname;
+  const isFetching = useSelector(isFetchingProductsList);
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(
@@ -47,29 +60,13 @@ function ProductsList({ perPage }) {
     );
   }, [dispatch, currentPage, categories, formattedMinPrice, formattedMaxPrice]);
 
-  useEffect(() => {
-    dispatch(getAllHomeProducts());
-  }, [dispatch]);
-
-  const countPagination = total ? Math.round(total / perPage) : 0;
-  const spacingHomePage = {
-    xs: 1.5,
-    sm: 2,
-    md: 2.5,
-    lg: 6,
-  };
-  const spacingProductsPage = {
-    xs: 1,
-    sm: 1.5,
-    md: 2,
-    lg: 3,
-  };
+  const countPagination = Math.round(total / perPage);
+  const location = useLocation();
+  const currentPath = location.pathname;
+  const spacingHomePage = { xs: 1.5, sm: 2, md: 2.5, lg: 6 };
+  const spacingProductsPage = { xs: 1, sm: 1.5, md: 2, lg: 3 };
   const gridSpacing =
     currentPath === '/' ? spacingHomePage : spacingProductsPage;
-  const isFetching =
-    currentPath === '/' ? isFetchingHomeProds : isFetchingProducts;
-  const filterProductsForHomePage = (productsForFilter) =>
-    productsForFilter.filter((product) => product.quantity !== 0);
   return (
     <div>
       {isFetching ? (
@@ -82,42 +79,35 @@ function ProductsList({ perPage }) {
         />
       ) : (
         <div>
-          <Grid
-            container
-            spacing={gridSpacing}
-            sx={{ padding: '0 1%', margin: '0 auto', width: '90%' }}>
-            {currentPath === '/'
-              ? filterProductsForHomePage(productsForHomePage).map(
-                  (product) => (
-                    <StyledGrid
-                      item
-                      xs={6}
-                      sm={6}
-                      md={4}
-                      lg={3}
-                      key={product.itemNo}
-                      sx={{ alignItems: 'baseline' }}>
-                      <ProductCard product={product} />
-                    </StyledGrid>
-                  )
-                )
-              : productsForProducts.map((product) => (
-                  <Grid
-                    display="flex"
-                    justifyContent="center"
-                    item
-                    xs={6}
-                    sm={6}
-                    md={6}
-                    lg={3}
-                    key={product.itemNo}
-                    height="auto"
-                    sx={{
-                      alignItems: 'baseline',
-                    }}>
-                    <ProductCard product={product} />
-                  </Grid>
-                ))}
+          <Grid container spacing={gridSpacing}>
+            {products.map((product) =>
+              currentPath === '/' ? (
+                <StyledGrid
+                  item
+                  xs={6}
+                  sm={6}
+                  md={4}
+                  lg={3}
+                  key={product.itemNo}
+                  sx={{ alignItems: 'baseline' }}>
+                  <ProductCard product={product} />
+                </StyledGrid>
+              ) : (
+                <Grid
+                  display="flex"
+                  justifyContent="center"
+                  item
+                  xs={6}
+                  sm={6}
+                  md={6}
+                  lg={3}
+                  key={product.itemNo}
+                  height="auto"
+                  sx={{ alignItems: 'baseline' }}>
+                  <ProductCard product={product} />
+                </Grid>
+              )
+            )}
           </Grid>
           {currentPath === '/' ? null : (
             <Box
