+++ client/src/pages/Products/index.jsx
@@ -5,11 +5,16 @@ import Filter from '../../components/Filter';
 
 function ProductsContent() {
   return (
-    <Stack
-      sx={{ width: '90%', margin: '0 auto' }}
-      direction={{ md: 'row', sm: 'column' }}>
+    <Stack direction={{ md: 'row', sm: 'column' }}>
       <Filter />
-      <Container>
+      <Container
+        sx={{
+          display: 'flex',
+          justifyContent: 'center',
+          flexDirection: 'column',
+          padding: '0 1%',
+          margin: '2% auto',
+        }}>
         <ProductsList />
       </Container>
     </Stack>
