+++ client/src/themes/themeFilter.js
@@ -4,7 +4,7 @@ const FilterStyles = styled('aside')({
   display: 'flex',
   flexDirection: 'column',
   borderRight: '1px solid #EBEBEB',
-  margin: '0 auto',
+  margin: '2% auto',
   paddingLeft: '10px',
 });
